Add automatic generation at config time.
The templates are expanded at configure time. It creates a new header
file that generates all the variables/functions needed. Adding this
during the autogen automagicaly generates for each of the available
modules the proper set of functions.

Reduce the amount of symbols defined.

Most of variables could be static inside functions, reducing the
impact in term of used symbols.

Signed-off-by: Clement Foyer <clement.foyer@inria.fr>